Split status box logic into readable components
