+++ src/components/account/statusBox/StatusBox.tsx
@@ -3,35 +3,10 @@ import { Message } from 'retranslate';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { StatusBoxRow } from './statusBoxRow/StatusBoxRow';
-import { usePendingApplications } from '../../common/apiHooks';
-import { Shimmer } from '../../common/shimmer/Shimmer';
-import { Application, ApplicationType } from '../../common/apiModels';
-
-interface UserConversion {
-  secondPillar: Conversion;
-  thirdPillar: Conversion;
-}
-
-interface Amount {
-  yearToDate: number;
-  total: number;
-}
-
-interface Conversion {
-  selectionComplete: boolean;
-  transfersComplete: boolean;
-  paymentComplete: boolean;
-  pendingWithdrawal: boolean;
-  contribution: Amount;
-  subtraction: Amount;
-}
-
-interface Fund {
-  fundManager: { name: string };
-  activeFund: boolean;
-  name: string;
-  pillar: number;
-}
+import SecondPillarStatusBox from './secondPillarStatusBox';
+import { Fund, UserConversion } from './types';
+import { StatusBoxLoader } from './StatusBoxLoader';
+import { StatusBoxTitle } from './StatusBoxTitle';
 
 interface StatusBoxType {
   memberNumber: number | null;
@@ -73,7 +48,7 @@ export const StatusBox: React.FunctionComponent<StatusBoxType> = ({
       <StatusBoxTitle />
 
       <div className="card card-secondary">
-        {renderSecondPillarFlow(conversion.secondPillar, secondPillarFunds, loading)}
+        <SecondPillarStatusBox />
 
         <StatusBoxRow
           ok={!payTuleva3}
@@ -113,110 +88,6 @@ export const StatusBox: React.FunctionComponent<StatusBoxType> = ({
     </>
   );
 };
-
-const StatusBoxLoader: React.FunctionComponent = () => {
-  return (
-    <>
-      <StatusBoxTitle />
-      <div className="card card-secondary">
-        <div className="d-flex p-2 status-box-row tv-table__row">
-          <Shimmer height={52} />
-        </div>
-        <div className="d-flex p-2 status-box-row tv-table__row">
-          <Shimmer height={52} />
-        </div>
-        <div className="d-flex p-2 status-box-row">
-          <Shimmer height={52} />
-        </div>
-      </div>
-      <div className="mt-4">
-        <Shimmer height={16} />
-      </div>
-    </>
-  );
-};
-
-const StatusBoxTitle: React.FunctionComponent = () => {
-  return (
-    <div className="row">
-      <div className="col-md-6 mb-4 lead">
-        <Message>account.status.choices</Message>
-      </div>
-    </div>
-  );
-};
-
-const renderSecondPillarFlow = (
-  secondPillar: Conversion,
-  secondPillarFunds: Fund[],
-  loading: boolean,
-) => {
-  const pendingWithdrawal = usePendingWithdrawalApplication();
-  if (secondPillar.pendingWithdrawal && pendingWithdrawal) {
-    return (
-      <StatusBoxRow
-        showAction={!loading}
-        name={<Message>account.status.choice.pillar.second</Message>}
-        lines={[<Message>account.status.choice.pillar.second.withdraw</Message>]}
-      >
-        <Link to={`/applications/${pendingWithdrawal.id}/cancellation`} className="btn btn-light">
-          <Message>account.status.choice.pillar.second.withdraw.cancel</Message>
-        </Link>
-      </StatusBoxRow>
-    );
-  }
-
-  const activeFund = secondPillarFunds.find((fund) => fund.activeFund);
-  if (!secondPillar.selectionComplete || !activeFund) {
-    const statusMessage = activeFund ? (
-      activeFund.name
-    ) : (
-      <Message>account.status.choice.pillar.second.missing</Message>
-    );
-
-    return (
-      <StatusBoxRow
-        showAction={!loading}
-        name={<Message>account.status.choice.pillar.second</Message>}
-        lines={[statusMessage]}
-      >
-        <Link to="/2nd-pillar-flow" className="btn btn-light">
-          <Message>account.status.choice.join.tuleva.2</Message>
-        </Link>
-      </StatusBoxRow>
-    );
-  }
-
-  if (!secondPillar.transfersComplete) {
-    return (
-      <StatusBoxRow
-        showAction={!loading}
-        name={<Message>account.status.choice.pillar.second</Message>}
-        lines={[<Message>account.status.choice.pillar.second.transferIncomplete</Message>]}
-      >
-        <Link to="/2nd-pillar-flow" className="btn btn-light">
-          <Message>account.status.choice.transfer.tuleva.2</Message>
-        </Link>
-      </StatusBoxRow>
-    );
-  }
-
-  return (
-    <StatusBoxRow
-      ok
-      name={<Message>account.status.choice.pillar.second</Message>}
-      showAction={!loading}
-      lines={[activeFund.name]}
-    />
-  );
-};
-
-const usePendingWithdrawalApplication = (): Application | undefined =>
-  usePendingApplications().data?.find(
-    (application) =>
-      application.type === ApplicationType.WITHDRAWAL ||
-      application.type === ApplicationType.EARLY_WITHDRAWAL,
-  );
 
 const mapStateToProps = (state: {
   login: {
