+++ src/components/account/statusBox/secondPillarStatusBox/index.js
@@ -1,0 +1,1 @@
+export { default } from './SecondPillarStatusBox';
